% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_vol.R
\name{media_vol}
\alias{media_vol}
\title{Calcular media do volume}
\usage{
media_vol(reservoir, year = 2022, month = 4)
}
\arguments{
\item{reservoir}{nome do reservatorio em texto}

\item{year}{ano em numero}

\item{month}{mes em numero}
}
\value{
uma tibble
}
\description{
Essa funcao calcula a media de volume (\%) armazenado ...
}
\examples{
media_vol("Cantareira")
}
