% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Partidas do Brasileirao}
\format{
\describe{
\item{temporada}{temporada em numero}
\item{data}{data do jogo}
\item{...}{...}
}
}
\source{
\url{https://github.com/williamorim/brasileirao}
}
\usage{
partidas_brasileirao
}
\description{
Essa base possui dados do Brasileirao, entre 2003 ate 2022.
}
\keyword{datasets}
